Cleanup homerow_mode.xml
Tried to place as many autogens as possible inside their items
instead of being all inside of blocks in one big item.

Exceptions to these are the autogens related to U/O Keys which proper
working depends on an order to be maintained.
The culprits of this complexity are 2 options related with U/O, which
now seem to be redundant considering the same results can be
achieved via combination of different and same number of keys.
Deleting these options has been considered however it is difficult to
determine the number of people which might already be using them.